requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
exceptiongroup==1.2.2; python_version < '3.11'
fastapi==0.112.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
idna==3.7; python_version >= '3.5'
jinja2==3.1.4; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
python-multipart==0.0.9; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn==0.30.6; python_version >= '3.8'
pymongo
python-dotenv
